+++ src/components/Header/index.tsx
@@ -1,5 +1,5 @@
-import styled from 'styled-components';
 import React from "react";
+import { Item, Link, NavigationBar } from './styles';
 
 const Header: React.FC = () => {
   return (
@@ -8,35 +8,5 @@ const Header: React.FC = () => {
     </NavigationBar>
   )
 }
-
-interface ActiveProps {
-  readonly isActive: boolean;
-}
-
-const NavigationBar = styled.ul`
-  list-style-type: none;
-  margin: 0;
-  padding: 0;
-  overflow: hidden;
-  background-color: #333;
-`;
-
-const Item = styled.li`
-  float: left;
-`;
-
-const Link = styled.a<ActiveProps>`
-  display: block;
-  color: white;
-  text-align: center;
-  padding: 14px 16px;
-  text-decoration: none;
-
-  &:hover {
-    background-color: #111;
-  }
-
-  background-color: ${(props) => (props.isActive ?  "#04AA6D" : null)};
-`;
 
 export default Header;
